feat: Card component and blog page card basic styling
