+++ components/BlockBuilder.vue
@@ -54,7 +54,7 @@ import ImageBlock from '@/components/ImageBlock'
 import VideoBlock from '@/components/VideoBlock'
 import AccordionBlock from '@/components/AccordionBlock'
 // import SliderBlock from '@/components/SliderBlock'
-// import CardListBlock from '@/components/CardListBlock'
+import CardListBlock from '@/components/CardListBlock'
 // import SocialIcons from '@/components/SocialIcons'
 // import TabbedSlider from '@/components/TabbedSlider'
 import BlockBuilder from '@/components/BlockBuilder'
@@ -71,7 +71,7 @@ export default {
     VideoBlock,
     AccordionBlock,
     // SliderBlock,
-    // CardListBlock,
+    CardListBlock,
     // SocialIcons,
     // TabbedSlider,
     BlockBuilder,
@@ -112,7 +112,7 @@ export default {
         case 'video_block' : name = 'VideoBlock'; break
         case 'accordion_block' : name = 'AccordionBlock'; break
         // case 'slider_block' : name = 'SliderBlock'; break
-        // case 'card_list_block' : name = 'CardListBlock'; break
+        case 'card_list_block' : name = 'CardListBlock'; break
         case 'custom' : name = block.component; break
       }
       return name
